Moved AdaptiveCardBuilder code
